4.21*: bloglist expansion, step9

+++ bloglist/controllers/blogs.js
@@ -54,8 +54,24 @@ blogsRouter.put('/:id', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blogToDelete = (await Blog.findById(request.params.id)).toJSON()
+
+  // Same lines as in the POST method
+  // I shouldn't repeat the code
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = (await User.findById(decodedToken.id)).toJSON()
+
+  if (blogToDelete.user.toString() !== user.id) {
+    // console.log(typeof blogToDelete.user)
+    // console.log(typeof user.id);
+    response.status(401).json({ error: 'only the author can delete a blog' })
+  } else {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).json({ success: 'blog removed successfully' })
+  }
 })
 
 module.exports = blogsRouter
